Move tests for provenance.py to tests-folder.
